requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
gunicorn>=22.0,<23.0
dj-database-url>=2.1,<3.0
whitenoise>=6.7,<7.0
